Created pages, sorted navbar on all, and responsiveness, cleaned up css
