--- src/components/About/About.jsx
@@ -1,13 +1,0 @@
-// About.js
-import React from "react";
-
-const About = () => {
-	return (
-		<div>
-			<h2>About</h2>
-			<p>This is the About page.</p>
-		</div>
-	);
-};
-
-export default About;
