+++ src/components/Input/Input.css
@@ -1,5 +1,5 @@
 .input-container {
-  width: clamp(200px, 50%, 600px);
+  width: clamp(300px, 50%, 600px);
   margin: 0 auto 30px auto;
   display: flex;
   justify-content: space-between;
@@ -18,10 +18,10 @@
   background-color: rgb(200, 200, 200);
 }
 .input-container input {
-  width: 75%;
+  width: 73%;
 }
 .input-container button {
-  width: 20%;
+  width: 25%;
 }
 .test-text-container {
   min-height: 100px;
