--- src/components/PlaceDetails/PlaceDetails.jsx
@@ -1,8 +1,0 @@
-import React from "react";
-import "./PlaceDetails.css";
-
-const PlaceDetails = () => {
-  return <h1>PlaceDetails</h1>;
-};
-
-export default PlaceDetails;
