+++ src/pages/About/About.jsx
@@ -1,0 +1,16 @@
+// About.jsx
+import React from "react";
+import "./About.css";
+import Nav from "../../components/Nav/Nav";
+
+const About = () => {
+  return (
+    <>
+      <Nav />
+      <h2>About</h2>
+      <p>This is the About page.</p>
+    </>
+  );
+};
+
+export default About;
