+++ src/pages/Contact/Contact.jsx
@@ -1,0 +1,16 @@
+// Contact.jsx
+import React from "react";
+import "./Contact.css";
+import Nav from "../../components/Nav/Nav";
+
+const Contact = () => {
+  return (
+    <>
+      <Nav />
+      <h2>Contact</h2>
+      <p>This is the Contact page.</p>
+    </>
+  );
+};
+
+export default Contact;
